CRUD functionality
Minor changes were made into the frontend. A big overhaul has been brought into the backend. Routes to Create, Read, Delete, Update functionality for the user account as well as the assignment functionality of the application has been defined.  Many of the part of the codes could be refined in the future as this is currently in development.